requirements.txt
Django==4.1.2
django-crispy-forms==1.14.0
djangorestframework==3.14.0